server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.11.0
httpx>=0.28.0
